Dev: qdevice: Split class QDevice into qdevice.py from corosync.py
